refactor: split graph module into focused files with barrel export
analyzer.ts -> analyzer.ts + tools.ts + labels.ts
render.ts -> render.ts + treemap.ts + sessions.ts
Added index.ts barrel for clean public API.
