+++ src/commands/graph/index.ts
@@ -5,7 +5,6 @@ import * as os from "node:os";
 import * as path from "node:path";
 import { x } from "tinyexec";
 import { BaseCommand } from "../base.js";
-import { parseJsonl } from "../../lib/graph/parser.js";
 import {
   buildAllSessionsTreemap,
   buildBarChartData,
@@ -13,27 +12,33 @@ import {
   findRecentSessions,
   findSessionPath,
   generateTreemapHtml,
-} from "../../lib/graph/render.js";
-import { openInBrowser, startServer, waitForShutdown } from "../../lib/graph/server.js";
+  openInBrowser,
+  parseJsonl,
+  startServer,
+  waitForShutdown,
+} from "../../lib/graph/index.js";
 
 // Re-export public API for consumers and tests
-export { calculateCutoffMs, filterByDays, parseJsonl } from "../../lib/graph/parser.js";
-export { analyzeSession, extractSessionLabel } from "../../lib/graph/analyzer.js";
 export {
-  buildBarChartData,
+  analyzeSession,
   buildAllSessionsTreemap,
+  buildBarChartData,
   buildSessionTreemap,
-  generateTreemapHtml,
+  calculateCutoffMs,
+  extractSessionLabel,
+  filterByDays,
   findRecentSessions,
   findSessionPath,
-} from "../../lib/graph/render.js";
+  generateTreemapHtml,
+  parseJsonl,
+} from "../../lib/graph/index.js";
 export type {
   BarChartData,
   BarChartDay,
   ProjectBar,
   SessionResult,
   TreemapNode,
-} from "../../lib/graph/types.js";
+} from "../../lib/graph/index.js";
 
 /**
  * Generate interactive HTML treemap from session token data
