+++ src/lib/graph/analyzer.test.ts
@@ -4,14 +4,16 @@ import {
   aggregateSessionTools,
   analyzeSession,
   extractProjectName,
-  extractSessionLabel,
-  extractToolData,
-  extractToolDetail,
   getModelName,
   getPrimaryModel,
+} from "./analyzer";
+import { extractSessionLabel } from "./labels";
+import {
+  extractToolData,
+  extractToolDetail,
   sanitizeString,
   truncateDetail,
-} from "./analyzer";
+} from "./tools";
 
 // ── Helpers ──
