+++ src/lib/graph/analyzer.ts
@@ -1,4 +1,5 @@
-import type { ContentBlock, JournalEntry, ToolData } from "./types.js";
+import type { JournalEntry, ToolData } from "./types.js";
+import { extractToolData } from "./tools.js";
 
 /**
  * Analyze session entries to get token breakdown by model.
@@ -68,183 +69,6 @@ export function analyzeSession(entries: JournalEntry[]): {
     repeatedReads,
     modelEfficiency: totalTokens > 0 ? opusTokens / totalTokens : 0,
   };
-}
-
-/**
- * Extract a meaningful label from session entries.
- */
-export function extractSessionLabel(entries: JournalEntry[], sessionId: string): string {
-  let firstUserText: string | undefined;
-  let firstAssistantText: string | undefined;
-  let gitBranch: string | undefined;
-  let slug: string | undefined;
-
-  for (const entry of entries) {
-    // Extract metadata from any entry
-    if (!gitBranch && (entry as any).gitBranch) {
-      gitBranch = (entry as any).gitBranch;
-    }
-    if (!slug && (entry as any).slug) {
-      slug = (entry as any).slug;
-    }
-
-    if (!entry.message) continue;
-
-    // Look for first user message with actual text (not UUID reference)
-    if (!firstUserText && entry.type === "user" && entry.message.role === "user") {
-      const content = entry.message.content;
-      if (typeof content === "string") {
-        // Check if it's a UUID (skip those) or actual text
-        const isUuid = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i.test(
-          content,
-        );
-        if (!isUuid && content.length > 0) {
-          firstUserText = content;
-        }
-      } else if (Array.isArray(content)) {
-        // Look for text blocks in array content
-        for (const block of content) {
-          if (block.type === "text" && block.text && block.text.length > 0) {
-            firstUserText = block.text;
-            break;
-          }
-        }
-      }
-    }
-
-    // Look for first assistant text response
-    if (!firstAssistantText && entry.type === "assistant" && entry.message.role === "assistant") {
-      const content = entry.message.content;
-      if (Array.isArray(content)) {
-        for (const block of content) {
-          if (block.type === "text" && block.text && block.text.length > 0) {
-            firstAssistantText = block.text;
-            break;
-          }
-        }
-      }
-    }
-
-    // Stop early if we have user text
-    if (firstUserText) break;
-  }
-
-  // Priority: user text > assistant text > git branch > slug > short ID
-  let label = firstUserText || firstAssistantText || gitBranch || slug || sessionId.slice(0, 8);
-
-  // Clean up the label
-  label = label
-    .replace(/^\/\S+\s*/, "") // Remove /command prefixes
-    .replace(/<[^>]+>[^<]*<\/[^>]+>/g, "") // Remove XML-style tags with content
-    .replace(/<[^>]+>/g, "") // Remove remaining XML tags
-    .replace(/^\s*Caveat:.*$/m, "") // Remove caveat lines
-    .replace(/\n.*/g, "") // Take only first line
-    // eslint-disable-next-line no-control-regex
-    .replace(/[\x00-\x1F]+/g, " ") // Replace control characters with space
-    .trim();
-
-  // If still empty or too short, use fallback
-  if (label.length < 3) {
-    label = slug || sessionId.slice(0, 8);
-  }
-
-  // Truncate very long labels (will be smart-truncated in UI based on box size)
-  if (label.length > 80) {
-    label = label.slice(0, 77) + "...";
-  }
-
-  return label;
-}
-
-/**
- * Sanitize string by replacing control characters (newlines, tabs, etc.) with spaces.
- */
-export function sanitizeString(str: string): string {
-  // Replace all control characters (ASCII 0-31) with space, collapse multiple spaces
-  // eslint-disable-next-line no-control-regex
-  return str.replace(/[\x00-\x1F]+/g, " ").trim();
-}
-
-/**
- * Truncate a string and extract just the filename for paths.
- */
-export function truncateDetail(str: string | undefined, maxLen = 30): string | undefined {
-  if (!str) return undefined;
-  // Sanitize control characters first
-  const sanitized = sanitizeString(str);
-  // For file paths, show just the filename
-  if (sanitized.includes("/")) {
-    const parts = sanitized.split("/");
-    const filename = parts[parts.length - 1];
-    return filename.length > maxLen ? filename.slice(0, maxLen - 3) + "..." : filename;
-  }
-  return sanitized.length > maxLen ? sanitized.slice(0, maxLen - 3) + "..." : sanitized;
-}
-
-/**
- * Extract a meaningful detail string from tool input.
- */
-export function extractToolDetail(
-  toolName: string,
-  input?: Record<string, unknown>,
-): string | undefined {
-  if (!input) return undefined;
-
-  switch (toolName) {
-    case "Read":
-      return truncateDetail(input.file_path as string);
-    case "Write":
-    case "Edit":
-      return truncateDetail(input.file_path as string);
-    case "Bash":
-      return truncateDetail(input.command as string, 50);
-    case "Glob":
-      return truncateDetail(input.pattern as string, 50);
-    case "Grep":
-      return truncateDetail(input.pattern as string, 50);
-    case "Task":
-      return truncateDetail(input.description as string, 50);
-    case "WebFetch":
-      return truncateDetail(input.url as string, 40);
-    default:
-      return undefined;
-  }
-}
-
-/**
- * Extract individual tool calls from message content blocks.
- * Returns each tool call with its detail (file path, command, etc.).
- */
-export function extractToolData(
-  content: ContentBlock[] | string | undefined,
-  turnInputTokens: number,
-  turnOutputTokens: number,
-): ToolData[] {
-  if (!content || typeof content === "string") return [];
-
-  // Collect individual tool_use blocks
-  const toolBlocks: Array<{ name: string; detail?: string }> = [];
-  for (const block of content) {
-    if (block.type === "tool_use" && block.name) {
-      const detail = extractToolDetail(block.name, block.input);
-      toolBlocks.push({ name: block.name, detail });
-    }
-  }
-
-  if (toolBlocks.length === 0) return [];
-
-  // Distribute tokens proportionally across individual calls
-  const tokensPerCall = {
-    input: Math.round(turnInputTokens / toolBlocks.length),
-    output: Math.round(turnOutputTokens / toolBlocks.length),
-  };
-
-  return toolBlocks.map((tool) => ({
-    name: tool.name,
-    detail: tool.detail,
-    inputTokens: tokensPerCall.input,
-    outputTokens: tokensPerCall.output,
-  }));
 }
 
 /**
