+++ src/lib/graph/index.ts
@@ -1,0 +1,20 @@
+// Barrel export for graph module public API
+
+export { analyzeSession, aggregateSessionTools, getPrimaryModel, getModelName, extractProjectName } from "./analyzer.js";
+export { extractSessionLabel } from "./labels.js";
+export { calculateCutoffMs, filterByDays, parseJsonl, quickTokenCount } from "./parser.js";
+export { generateTreemapHtml } from "./render.js";
+export { openInBrowser, startServer, waitForShutdown } from "./server.js";
+export { findRecentSessions, findSessionPath, buildBarChartData } from "./sessions.js";
+export { sanitizeString, truncateDetail, extractToolDetail, extractToolData } from "./tools.js";
+export { buildTurnNodes, buildSessionTreemap, buildAllSessionsTreemap } from "./treemap.js";
+export type {
+  BarChartData,
+  BarChartDay,
+  ContentBlock,
+  JournalEntry,
+  ProjectBar,
+  SessionResult,
+  ToolData,
+  TreemapNode,
+} from "./types.js";
